+++ R/data.R
@@ -68,7 +68,7 @@
 #' }
 "ExPanD_config_russell_3000"
 
-#' A Snapshot of Macroeconomic Data as Provided by the World Bank API (1960 - 2016)
+#' A Snapshot of Macroeconomic Data as Provided by the World Bank API (1960 - 2018)
 #'
 #' Data collected from the World Bank API using the package \code{wbstats}.
 #'
@@ -81,7 +81,7 @@
 #' @keywords data datasets macroeconomic
 #'
 #' @source Has been collected using the \code{\link[wbstats]{wb}} function
-#'  from the World Bank API in March 2018.
+#'  from the World Bank API in Jan 2020.
 #'  The code to generate this data is available in the
 #'  \href{https://github.com/joachim-gassen/ExPanDaR}{github repository} of this package.
 #'  Use in scientific studies is not advised without prior cleaning/checking.