+++ R/prepare_quantile_trend_graph.R
@@ -13,6 +13,7 @@
 #' @param var a string containing the column name of the variable
 #'   to be plotted. Defaults to the last numerical variable of the data frame
 #'   that is not \code{ts_id}.
+#' @param points Do you want points to indicate the statistics? Defaults to \code{TRUE}.
 #'
 #' @return A list containing two items:
 #' \describe{
@@ -31,7 +32,8 @@
 
 prepare_quantile_trend_graph <- function(df, ts_id,
                                          quantiles = c(0.05, 0.25, 0.5, 0.75, 0.95),
-                                         var = utils::tail(colnames(df[sapply(df, is.numeric) & colnames(df) != ts_id]), n = 1)) {
+                                         var = utils::tail(colnames(df[sapply(df, is.numeric) & colnames(df) != ts_id]), n = 1),
+                                         points = TRUE) {
   if(!is.data.frame(df)) stop("df needs to be a dataframe")
   df <- as.data.frame(df)
   if (! ts_id %in% colnames(df)) stop("ts_id need to be in df")
@@ -39,16 +41,12 @@ prepare_quantile_trend_graph <- function(df, ts_id,
   if (! var %in% colnames(df)) stop("var needs to be in df")
 
   df <- droplevels(df[stats::complete.cases(df[, c(ts_id, var)]), c(ts_id, var)])
-  if(!is.numeric(ts_id)) xnum <- suppressWarnings(as.numeric(as.character(df[,ts_id]))) else xnum <- ts_id
-  if (anyNA(xnum)) {
-    x_is_factor <- TRUE
-  } else {
-    df[,ts_id] <- xnum
-    x_is_factor <- FALSE
-  }
+  df[, ts_id] <- try_convert_ts_id(df[, ts_id])
+
   q <- sapply(quantiles,
               function(x){by(df[, which(!colnames(df) %in% ts_id)], df[, ts_id], stats::quantile, na.rm=TRUE,x)})
-  if (x_is_factor) q <- data.frame(rownames(q),q) else q <- data.frame(as.numeric(rownames(q)),q)
+  q <- data.frame(rownames(q),q)
+  q[1] <- sort(unique(df[, ts_id]))
   colnames(q)[1] <- ts_id
   colnames(q)[2:(length(quantiles) + 1)] <- sprintf("q%02d", quantiles*100)
   gg <- tidyr::gather_(data = q,
@@ -56,15 +54,18 @@ prepare_quantile_trend_graph <- function(df, ts_id,
                        value_col = colnames(df)[!(ts_id == colnames(df))],
                        gather_cols = colnames(q)[!(ts_id == colnames(q))])
   gg$quantile <- factor(gg$quantile, levels = unique(gg$quantile))
-  if (x_is_factor) {
+  if (is.factor(df[, ts_id])) {
     plot <- ggplot2::ggplot(gg, ggplot2::aes_string(x = ts_id, y=gg[,3], color="quantile", group = "quantile")) +
-      ggplot2::stat_summary(fun.y=sum, geom="line") + ggplot2::ylab(colnames(df[ncol(df)])) +
+      ggplot2::stat_summary(fun = sum, geom = "line") +
+      ggplot2::ylab(colnames(df[ncol(df)])) +
       ggplot2::scale_color_discrete(labels=quantiles)
   } else {
-    plot <- ggplot2::ggplot(gg) + ggplot2::ylab(colnames(df[ncol(df)])) +
-      ggplot2::geom_line(ggplot2::aes_string(x=ts_id, y=gg[,3], color="quantile")) +
+    plot <- ggplot2::ggplot(gg, ggplot2::aes_string(x=ts_id, y=gg[,3], color="quantile")) +
+      ggplot2::ylab(colnames(df[ncol(df)])) +
+      ggplot2::geom_line() +
       ggplot2::scale_color_discrete(labels=quantiles)
   }
+  if (points) plot <- plot + ggplot2::geom_point()
   list(df = gg, plot = plot)
 }